:dizzy:[update]move get_closer_distance() to/sphere
