+++ includes/object.h
@@ -55,7 +55,6 @@ double			get_distance(const t_ray *ray, void *object);
 // object/distance.c
 void			calc_distance_by_discriminant(\
 	const double a, const double b, const double d, double *distance);
-double			get_closer_distance(double discriminant, double *distance);
 
 // object/intersection.c
 bool			is_shadow_intersection(\
@@ -69,11 +68,12 @@ t_sphere		*init_sphere(const char **line);
 // sphere/sphere_ray.c
 double			calc_discriminant_for_sphere(\
 	const t_ray *ray, const t_sphere *sphere, double *distances);
-double			get_distance_to_sphere(\
-	const t_ray *ray, t_sphere *sphere);
 
 // sphere/sphere_color.c
 t_vector		get_normal_on_sphere(t_intersection intersection, const t_ray *ray);
+
+// sphere/sphere_distance.c
+double			get_distance_to_sphere(const t_ray *ray, t_sphere *sphere);
 
 // plane/plane.c
 t_plane			*init_plane(const char **line);
